Make terraform devtools work with `.terraform` as a volume
Stuff that goes into .terraform is arch and OS dependent so it is best
that this directory not be shared between host and guest, however if
you mounted a volume to `.terraform` then the Makefile did not do
terraform initialization.

This patch changes the Makefile to also check for
`.terraform/providers` and do initialization if it does not exist.
This should work in all cases where there at least one provider.

Also:
- Simplified the testing of the terraform image
- Enable terraform caching and test with and without caching