Add socat and busybox to the techdocs image
The live preview of techdocs listens on `127.0.0.1`, and if this is inside
a docker container it is not possible to get to it via port bindings
[[ref](https://stackoverflow.com/a/52518929)].

`socat` can be used to re-expose the stuff listening on `127.0.0.1`.

I added busybox mainly because it provides a bunch of useful tools that
can be used to debug various issues with the image.