feat: beefier+opinionated poetry python3.10 devtools
Added new targets to poetry-python3.10

- poetry-install
- validate
- validate-static
- validate-fix
- test
- test-verbose

This to to migrate `offers-data`
to standard toolings and is an opinionated implementation
of what is specifically needed for that project. The makefile targets
introduced are simply "copy" of targets from `offers-data` project.

The idea is to implement something and improve it in future
when other projects start using this tool.