Support appending to the supplied Dockerfile in devtools-golang
The Golang devtools support the building of an OCI either from a
devtool-supplied Dockerfile, or when customizations are required, a
user-supplied Dockerfile.

To avoid unnecessary duplication of the supplied Dockerfile in the case
where only a few additional Dockerfile commands are needed, add support
for modifying it via an append.