Support copying of additional resource files in devtool-supplied OCI image
The Golang devtools support the building of an OCI image for an
application either from a devtool-supplied Dockerfile, or when
significant customizations are required, a user-supplied Dockerfile.

To avoid unnecessary duplication of the supplied Dockerfile in the case
where only a few additional resource files are needed, add support
for copying a list of paths into the application's working directory.