chore(deps): Upgrade to golangci-lint v2
This change upgrades Golangci-lint to version 2. With the upgrade brings
changes to several linters ([Full list of changes])

The most notable is the merge of the linters `stylecheck`, `gosimple`,
and `staticcheck` into `staticcheck`, the 3 linters all implement
https://staticcheck.dev/ rules.

The change causes the default rules from `stylecheck` to be active.

https://github.com/coopnorge/mage uses golangci-lint v2 with default
settings for static check. Initial testing of mage has detected violations of
[ST1005]: Incorrectly formatted error string. Further testing might
bring other violations to the surface.

The configuration was migrated using:

 ```console
golangci-lint migrate
```

Closes: https://github.com/coopnorge/engineering-issues/issues/444

[Full list of of changes]: https://golangci-lint.run/product/migration-guide/#changes