feat(TechDocs): Let Prettier be the authorative decider for certain markdown decisions
This PR promotes the `format-check` and `format-fix` of the TechDocs image to be included in the `validate` and `validate-fix` targets.

It also disables the Markdownlint-rules `# MD013 line-length - Line length` and `MD030 list-marker-space - Spaces after list markers`, since Prettier will now deal with these.

These two specific rules have had minor conflicts between the formatter and the linter, but since these are only related to formatting, the formatter should be the authorative source for these decisions.